Updated Berksfile and MapR repositories
This commit updates the Berksfile and MapR metadata to add the
yum-epel cookbook, as the epel package was split off from the
main yum cookbook.

The MapR repos.rb recipe was updated to use the gpgkey specified in
either cluster.json, or in the attribute/default.rb. Previously,
this value was only used for debian OSes, and hardcoded for all
others. The default repository gpgkey url was also updated to the
more recent gpgkey