Integrated Hive, Hive Metastore, and HiveServer2 into Cookbook
Previously, the recipe for Hive existed, but was not actually used.
I've integrated it, and the metastore and server into the
mapr_control_node and mapr_data_node roles.

It successfully installs the components, but there is still some
configuration issue somewhere that exists even without these
changes and prevents login to the MapR Web UI.