Fixed MapR MCS Login Failure
The MapR MCS Webserver was not allowing a user with a valid name
and password to login. This is due to the webserver startup script
not properly sourcing the MapR env.sh in conf.new and therefore lacking
a necessary java configuration file.

An additional recipe and template file that sources env.sh fixed this.